Move part sorting to merge
