+++ client/js/libs/handlebars/ircmessageparser/merge.js
@@ -13,7 +13,11 @@ function assign(textPart, fragment) {
 	return Object.assign({}, fragment, {start, end, text});
 }
 
-// Merge the style fragments withing the text parts, taking into account
+function sortParts(a, b) {
+	return a.start - b.start || b.end - a.end;
+}
+
+// Merge the style fragments within the text parts, taking into account
 // boundaries and text sections that have not matched to links or channels.
 // For example, given a string "foobar" where "foo" and "bar" have been
 // identified as parts (channels, links, etc.) and "fo", "ob" and "ar" have 3
@@ -28,8 +32,18 @@ function merge(textParts, styleFragments) {
 	// is filled with "text" parts, dummy objects with start/end but no extra
 	// metadata.
 	const allParts = textParts
+		.sort(sortParts) // Sort all parts identified based on their position in the original text
 		.concat(fill(textParts, cleanText))
-		.sort((a, b) => a.start - b.start);
+		.sort(sortParts) // Sort them again after filling in unstyled text
+		.reduce((prev, curr) => {
+			const intersection = prev.some((p) => anyIntersection(p, curr));
+
+			if (intersection) {
+				return prev;
+			}
+
+			return prev.concat([curr]);
+		}, []);
 
 	// Distribute the style fragments within the text parts
 	return allParts.map((textPart) => {
